requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
whitenoise==6.8.2